backend/simulators/simulator_1/backend/requirements.txt
sqlalchemy==1.4.21
fastapi==0.68.2
pydantic==1.9.0
uvicorn==0.15.0
psycopg2-binary>=2.9.3,<3.0.0
pandas==1.4.1
requests==2.27.1
SQLAlchemy==1.4.21
pyarrow==7.0.0